+++ src/main/java/de/maxhenkel/delivery/tasks/Group.java
@@ -63,6 +63,7 @@ public class Group implements INBTSerializable<CompoundNBT> {
         this();
         this.name = name;
         this.password = password;
+        initInitialInbox();
     }
 
     public Group() {
@@ -197,6 +198,13 @@ public class Group implements INBTSerializable<CompoundNBT> {
     public void onReachComputerAge() {
         ItemStack parcel = new ItemStack(ModItems.SEALED_PARCEL);
         ModItems.SEALED_PARCEL.setContents(parcel, NonNullList.from(ItemStack.EMPTY, new ItemStack(ModBlocks.COMPUTER)));
+        ModItems.SEALED_PARCEL.setSender(parcel, new TranslationTextComponent("message.delivery.unknown"));
+        addItemToInbox(parcel);
+    }
+
+    public void initInitialInbox() {
+        ItemStack parcel = new ItemStack(ModItems.SEALED_PARCEL);
+        ModItems.SEALED_PARCEL.setContents(parcel, NonNullList.from(ItemStack.EMPTY, new ItemStack(ModItems.ENVELOPE, 8), new ItemStack(ModItems.PARCEL, 4)));
         ModItems.SEALED_PARCEL.setSender(parcel, new TranslationTextComponent("message.delivery.unknown"));
         addItemToInbox(parcel);
     }